[FE] refactor : review 도메인 API 호출 부분 정리
