+++ frontend/src/service/@shared/api/index.ts
@@ -1,0 +1,1 @@
+export * as reviewAPI from './review';
